+++ R/balanceTest.R
@@ -250,6 +250,9 @@ balanceTest <- function(fmla,
   ## we wanted to allow departures from the ETT default.
   ## Something like `design@Sweights <- DesignWeights(aggDesign, <...>)`.)
   descriptives    <- designToDescriptives(design, covariate.scales)
+  ## The notmissing indicators can't themselves be missing, so no role for
+  ## another notmissing indicator to record where its missingness pattern.
+  ## Instead we just record a blank for them:
   NMpatterns <- c(NMpatterns, rep("", dim(descriptives)[1]-length(NMpatterns)))
 
   # these weights govern inferential but not descriptive calculations
@@ -280,7 +283,19 @@ balanceTest <- function(fmla,
 
   dimnames(descriptives)[[2]][nstats.previous + 1:2] <- c("z", "p")
 
-  # strip out summaries of not-missing indicators that only ever take the value T
+  inferentials <- do.call(rbind, lapply(tmp, function(s) {
+    data.frame(s$Msq, s$DF, pchisq(s$Msq, df = s$DF, lower.tail = FALSE))
+  }))
+  colnames(inferentials) <- c("chisquare", "df", "p.value")
+
+  tcovs <- lapply(tmp, function(r) {
+    tcov <- r$tcov
+    dimnames(tcov) <-
+      rep(list(dimnames(descriptives)[["vars"]]),2)
+    tcov
+  })
+
+  # Deal with summaries of not-missing indicators that only ever take the value T
   nmvars <- identify_NM_vars(dimnames(descriptives)[["vars"]])
   # next line assumes every "stat" not in the given list is a mean
   # over a group assigned to some treatment condition.
@@ -293,35 +308,39 @@ balanceTest <- function(fmla,
 	  toremove <- match(nmvars[bad], dimnames(descriptives)[["vars"]])
 	  if(length(toremove)>0){ ## if toremove=integer(0) then it drops all vars from descriptives
 		  descriptives <- descriptives[-toremove,,,drop=FALSE]
-		  origvars <- origvars[-toremove]
-                  strings_to_remove <- dimnames(descriptives)[["vars"]][toremove]
-                  NMpatterns <- NMpatterns[-toremove]  # names of vars that
-                  NMpatterns[ NMpatterns%in% strings_to_remove] <- ""
+	    tcovs <- lapply(tcovs,
+        function(mat){mat[-toremove, -toremove, drop=FALSE]}
+        )
+      origvars <- origvars[-toremove]
+
+      ## also trim covars to NM patterns lookup table:
+      NMpatterns <- NMpatterns[-toremove]
+      ## If the NM pattern var that a covar got mapped to
+      ## was removed, we take the covar to have no missings
+      ## anywhere, and report "" as the name of its nonmissing
+      ## indicator variable:
+      strings_to_remove <- dimnames(descriptives)[["vars"]][toremove]
+      NMpatterns[ NMpatterns%in% strings_to_remove] <- ""
 	  }
   }
 
-  inferentials <- do.call(rbind, lapply(tmp, function(s) {
-    data.frame(s$Msq, s$DF, pchisq(s$Msq, df = s$DF, lower.tail = FALSE))
-  }))
-  colnames(inferentials) <- c("chisquare", "df", "p.value")
+  for (s in 1L:dim(descriptives)[3]) {
+  descriptives[, "p", s] <- 
+    p.adjust(descriptives[, "p", s], method = p.adjust.method)
+  }
+  attr(descriptives, "NMpatterns") <- NMpatterns
+  attr(descriptives, "originals") <- origvars
+  attr(descriptives, "term.labels") <- design@TermLabels
+# Now `attr(d, "term.labels")[attr(d, "originals")]`
+# associates variables to terms of fmla
+  attr(descriptives, "include.NA.flags") <- 
+    include.NA.flags # hint to print and plot methods
 
-  # the meat of our xbal object
+# the meat of our xbal object
   ans$overall <- inferentials
+  attr(ans$overall, "tcov") <- tcovs
   ans$results <- descriptives
 
-  ## do p.value adjustment
-  for (s in 1L:dim(descriptives)[3])
-  ans$results[, "p", s] <- p.adjust(ans$results[, "p", s], method = p.adjust.method)
-##  ans$overall[, "p.value"] <- p.adjust(ans$overall[, "p.value"], method = p.adjust.method)
-
-  attr(ans$results, "NMpatterns") <- NMpatterns
-  attr(ans$results, "originals") <- origvars
-  attr(ans$results, "term.labels") <- design@TermLabels
-  attr(ans$results, "include.NA.flags") <- include.NA.flags # hinting for print and plot methods
-
-  attr(ans$overall, "tcov") <- lapply(tmp, function(r) {
-    r$tcov
-  })
   attr(ans, "fmla") <- formula(fmla)
   attr(ans, "report") <- report # hinting to our summary method later
   class(ans) <- c("balancetest", "xbal", "list")